Deprecate scrapy.utils.versions.
