+++ tests/ignores.txt
@@ -5,3 +5,4 @@ scrapy/mail.py
 scrapy/utils/console.py
 scrapy/utils/gz.py
 scrapy/utils/job.py
+scrapy/utils/versions.py
